+++ src/app/inobeta-ui/ui/material-forms/material-form/material-form.component.spec.ts
@@ -23,7 +23,7 @@ import { MatLegacyTooltipModule as MatTooltipModule } from '@angular/material/le
 import { Component } from '@angular/core';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { By } from '@angular/platform-browser';
-import { FlexLayoutModule } from '@angular/flex-layout';
+import { FlexLayoutModule } from '@Inobeta/flex-layout';
 import { IbMatTextboxControl, IbMatTextboxComponent } from '../controls/textbox';
 import { IbMatCheckboxControl, IbMatCheckboxComponent } from '../controls/checkbox';
 import { IbMatRadioControl, IbMatRadioComponent } from '../controls/radio';